Move options and add help
