+++ cmds/ocm/commands/ocmcmds/references/add/cmd.go
@@ -1,12 +1,12 @@
 package add
 
 import (
+	"github.com/open-component-model/ocm/cmds/ocm/commands/ocmcmds/common/addhdlrs"
 	"github.com/spf13/cobra"
 
 	"github.com/open-component-model/ocm/cmds/ocm/commands/ocmcmds/common"
 	"github.com/open-component-model/ocm/cmds/ocm/commands/ocmcmds/common/addhdlrs/refs"
 	"github.com/open-component-model/ocm/cmds/ocm/commands/ocmcmds/names"
-	"github.com/open-component-model/ocm/cmds/ocm/commands/ocmcmds/resources/add"
 	"github.com/open-component-model/ocm/cmds/ocm/commands/verbs"
 	"github.com/open-component-model/ocm/cmds/ocm/pkg/utils"
 	"github.com/open-component-model/ocm/pkg/contexts/clictx"
@@ -26,7 +26,7 @@ type Command struct {
 func NewCommand(ctx clictx.Context, names ...string) *cobra.Command {
 	return utils.SetupCommand(
 		&Command{
-			common.NewResourceAdderCommand(ctx, refs.New().WithCLIOptions(&add.Options{}), NewReferenceSpecificatonProvider()),
+			common.NewResourceAdderCommand(ctx, refs.New().WithCLIOptions(&addhdlrs.Options{}), NewReferenceSpecificatonProvider()),
 		},
 		utils.Names(Names, names...)...,
 	)
