+++ cmds/ocm/commands/ocmcmds/resources/add/cmd.go
@@ -1,6 +1,7 @@
 package add
 
 import (
+	"github.com/open-component-model/ocm/cmds/ocm/commands/ocmcmds/common/addhdlrs"
 	"github.com/spf13/cobra"
 
 	"github.com/open-component-model/ocm/cmds/ocm/commands/ocmcmds/common"
@@ -27,7 +28,7 @@ type Command struct {
 func NewCommand(ctx clictx.Context, names ...string) *cobra.Command {
 	return utils.SetupCommand(
 		&Command{
-			common.NewResourceAdderCommand(ctx, rscs.New().WithCLIOptions(&Options{}), NewResourceSpecificationsProvider(ctx, "")),
+			common.NewResourceAdderCommand(ctx, rscs.New().WithCLIOptions(&addhdlrs.Options{}), NewResourceSpecificationsProvider(ctx, "")),
 		},
 		utils.Names(Names, names...)...,
 	)
@@ -66,8 +67,8 @@ $ ocm add resources --file path/to/ca  resources.yaml VERSION=1.0.0
 
 func (o *Command) Long() string {
 	return `
-Add resources specified in a resource file to a component version.
-So far only component archives are supported as target.
+Adds resources specified in a resource file to a component version.
+So far, only component archives are supported as target.
 
 This command accepts resource specification files describing the resources
 to add to a component version. Elements must follow the resource meta data
